+++ src/main/java/ru/mipt/bit/platformer/game/LevelListenerGraphics.java
@@ -1,7 +1,7 @@
 package ru.mipt.bit.platformer.game;
 
 import ru.mipt.bit.platformer.graphics.GraphicsHandler;
-import ru.mipt.bit.platformer.graphics.Renderable;
+import ru.mipt.bit.platformer.graphics.objects.Renderable;
 import ru.mipt.bit.platformer.graphics.RenderableState;
 
 public class LevelListenerGraphics implements LevelListener {
@@ -14,14 +14,27 @@ public class LevelListenerGraphics implements LevelListener {
     @Override
     public void add(GameObject gameObject) {
         if (gameObject instanceof Renderable) {
-            graphicsHandler.addGraphicsObjects((Renderable) gameObject, RenderableState.ACTIVE);
+            graphicsHandler.addGraphicsObjects((Renderable) gameObject, convertState(gameObject.getState()));
         }
     }
 
     @Override
     public void parseState(GameObject gameObject) {
         if (gameObject instanceof Renderable) {
-            graphicsHandler.parseState((Renderable) gameObject, RenderableState.ACTIVE);
+            graphicsHandler.parseState((Renderable) gameObject, convertState(gameObject.getState()));
+        }
+    }
+
+    private RenderableState convertState(GameObjectState state) {
+        switch (state) {
+            case ALIVE: {
+                return RenderableState.ACTIVE;
+            }
+            case DEAD: {
+                return RenderableState.INACTIVE;
+            }
+            default:
+                throw new IllegalArgumentException("Unknown GameObjectState");
         }
     }
 }
