+++ src/main/java/ru/mipt/bit/platformer/graphics/GraphicsHandler.java
@@ -6,17 +6,21 @@ import com.badlogic.gdx.maps.MapRenderer;
 import com.badlogic.gdx.maps.tiled.TiledMap;
 import com.badlogic.gdx.maps.tiled.TiledMapTileLayer;
 import com.badlogic.gdx.maps.tiled.TmxMapLoader;
+import com.badlogic.gdx.math.GridPoint2;
 import com.badlogic.gdx.math.Interpolation;
-import ru.mipt.bit.platformer.game.GameObjectState;
-import ru.mipt.bit.platformer.graphics.objects.GameObjectGraphics;
-import ru.mipt.bit.platformer.graphics.objects.HealthBarGraphics;
-import ru.mipt.bit.platformer.graphics.util.TileMovement;
+import org.lwjgl.opengl.GLXSGIMakeCurrentRead;
+import ru.mipt.bit.platformer.game.objectsWithHelpers.objects.tank.Tank;
+import ru.mipt.bit.platformer.graphics.objects.*;
+import ru.mipt.bit.platformer.game.util.TileMovement;
 import ru.mipt.bit.platformer.game.objectsWithHelpers.objects.tank.Hittable;
 
+import java.lang.reflect.Constructor;
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
 import java.util.*;
 
-import static ru.mipt.bit.platformer.graphics.util.GdxGameUtils.createSingleLayerMapRenderer;
-import static ru.mipt.bit.platformer.graphics.util.GdxGameUtils.getSingleLayer;
+import static ru.mipt.bit.platformer.game.util.GdxGameUtils.createSingleLayerMapRenderer;
+import static ru.mipt.bit.platformer.game.util.GdxGameUtils.getSingleLayer;
 
 public class GraphicsHandler implements Graphics, Toggleable{
     private final Batch batch;
@@ -27,15 +31,16 @@ public class GraphicsHandler implements Graphics, Toggleable{
     private final Map<Renderable, Graphics> objectToGraphics = new HashMap<>();
     private final Collection<Graphics> graphicsObjects = new ArrayList<>();
     //TODO: rename
-    private final Map<Map.Entry<Class<?>, RenderableState>, String> classStateToPath;
+    private final Map<Map.Entry<Class<?>, RenderableState>, String> classStateToTexturePath;
+    private Map<Class<?>, HashSet<Class<? extends Graphics>>> classToGraphics = new HashMap<>();
 
-    public GraphicsHandler(String pathGameField, Map<Map.Entry<Class<?>, RenderableState>, String> classStateToPath) {
+    public GraphicsHandler(String pathGameField, Map<Map.Entry<Class<?>, RenderableState>, String> classStateToTexturePath){
         this.batch = new SpriteBatch();
         this.level = new TmxMapLoader().load(pathGameField);
         this.levelRenderer = createSingleLayerMapRenderer(level, batch);
         this.groundLayer = getSingleLayer(level);
         this.tileMovement = new TileMovement(groundLayer, Interpolation.smooth);
-        this.classStateToPath = classStateToPath;
+        this.classStateToTexturePath = classStateToTexturePath;
     }
 
     public void render() {
@@ -56,7 +61,7 @@ public class GraphicsHandler implements Graphics, Toggleable{
     }
 
     public void addGraphicsObjects(Renderable renderable, RenderableState state) {
-        String path = classStateToPath.get(new AbstractMap.SimpleEntry<>(renderable.getClass(), state));
+        String path = classStateToTexturePath.get(new AbstractMap.SimpleEntry<>(renderable.getClass(), state));
         if (path != null) {
             Graphics graphics = new GameObjectGraphics(path, renderable, tileMovement, batch);
             graphicsObjects.add(graphics);
