+++ src/main/java/ru/mipt/bit/platformer/graphics/objects/GameObjectGraphics.java
@@ -4,12 +4,10 @@ import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.Batch;
 import com.badlogic.gdx.graphics.g2d.TextureRegion;
 import com.badlogic.gdx.math.Rectangle;
-import ru.mipt.bit.platformer.graphics.Graphics;
-import ru.mipt.bit.platformer.graphics.Renderable;
-import ru.mipt.bit.platformer.graphics.util.TileMovement;
+import ru.mipt.bit.platformer.game.util.TileMovement;
 
-import static ru.mipt.bit.platformer.graphics.util.GdxGameUtils.createBoundingRectangle;
-import static ru.mipt.bit.platformer.graphics.util.GdxGameUtils.drawTextureRegionUnscaled;
+import static ru.mipt.bit.platformer.game.util.GdxGameUtils.createBoundingRectangle;
+import static ru.mipt.bit.platformer.game.util.GdxGameUtils.drawTextureRegionUnscaled;
 
 public class GameObjectGraphics implements Graphics {
     private final TextureRegion graphics;
