+++ src/main/java/ru/mipt/bit/platformer/graphics/objects/Toggleable.java
@@ -1,4 +1,4 @@
-package ru.mipt.bit.platformer.graphics;
+package ru.mipt.bit.platformer.graphics.objects;
 
 public interface Toggleable {
     void toggle();
